+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to Semantic Versioning ([SemVer]).
 ### Changes
 
 - Moved state directory from the service user account's home directory to `/var/lib/resonite-headless`. [@Cyberboss]
+- Moved cache directory to `/var/cache/resonite-headless`. [@Cyberboss]
 
 ### Fixed